Fix a max-depth-related bug in TreeRankNode.
For the history reasons: it could be that ranks[0] != maxRank
if the new-becoming "weak" was built under Split.NULL_MAX_DEPTH
on overflow of its new parent. In this case, maxRank was correct
but ranks[0] was not, in which case a wrong rank was returned
in a subsequent dominance check.