Implement threshold tuning in the ENS hybrid.
The version with tuning is not the default one yet;
it is available under the separate "jfb.rbtree.hybrid.ens.tune" ID.
As of now, the tuning is not yet perfect, as our estimations
of the running time of the JFB core are not ideal.
With good hyper-parameter values, it is able to yield a 20-30% speedup.